strafe and advance unit tests
